Change scaling syntax from = back to +
