+++ experiments/hpl/experiment.py
@@ -7,7 +7,7 @@ from benchpark.directives import variant, maintainers
 from benchpark.experiment import Experiment
 from benchpark.mpi import MpiOnlyExperiment
 from benchpark.openmp import OpenMPExperiment
-from benchpark.new_scaling import ScalingMode, Scaling
+from benchpark.scaling import ScalingMode, Scaling
 from benchpark.caliper import Caliper
