+++ lib/benchpark/experiment.py
@@ -419,7 +419,7 @@ class Experiment(ExperimentSystemBase, ExecMode, Affinity, Hwloc):
 
         self.compute_applications_section()
 
-        if "scaling" in self.spec.variants and not self.spec.satisfies("scaling=off"):
+        if any([self.spec.satisfies(s) for s in ["+strong", "+weak", "+throughput"]]):
             self.expr_vars.extend(self.scale())
 
         for var in self.expr_vars.values():
